Transports: a vessel is a map, and the deck is where her life is kept
Ported from 01, which is where these fixes were developed, and adapted to 5.4.8
rather than copied. Built on FreeBSD 15 / clang 19, ninja, WITHOUT the script
libraries (-DSCRIPT_LIB_ELUNA=0 -DSCRIPT_LIB_SD3=0, so MANGOS_SCRIPT_COMPAT is
undefined and the compiler is the check): 390/390, mangosd linked, zero errors,
mangos_tests 142 passed / 0 failed.

THE RIVAL IMPLEMENTATION IS GONE, and it was not merely residue. Pet::
MoveTransportFollow, SendSplineAnchor, UpdateTransport and Pet's own m_transport
flag did three forbidden things at once:

  * composed hull_world (X) local_offset and filed the pet into a grid cell from
    the result. A MO_TRANSPORT's server pose is a `GameTime % period` waypoint
    estimate; the client interpolates the real hull from TransportAnimation.dbc,
    which this server never loads. The pet was placed against a ship that is not
    there.
  * hand-built SMSG_MONSTER_MOVE with a transport guid, bypassing MoveSplineInit
    entirely, so movespline and the wire described different objects -- then
    fought the movement system with DisableSpline() + MotionMaster::Clear().
  * kept a SECOND notion of being aboard beside WorldObject::m_transportInfo, and
    pushed create/destroy for pets at boarding outside the visibility sweep. That
    is the ghosting: the server's record and the client's copy disagreeing.

None of it is needed once the deck is a map: the pet stands on it, DrawMinionTo
moves it between maps with ForgetMinion keeping m_clientGUIDs honest, and
MoveFollow routes on the deck's own navmesh in deck-local coordinates.

VISIBILITY ON BOARDING AND STEPPING ASHORE is three pieces, not one:

  * Player::SetCrossingVessel wraps the Remove/Add pair, and Map::Update-
    ObjectVisibility returns early while it is set. Mid-crossing he is off both
    maps for an instant, and out of world means invisible -- notifying then
    DESTROYS him for everyone about to be handed him back unchanged. That is
    "the man on the pier watched me vanish when I stepped aboard and reappear
    only once I moved". No create/destroy blocks ride the map swap; the change
    of transport rides the movement broadcast.
  * the object-side relay: a deck map also visits the world map's cells around
    the vessel, PASSENGERS ONLY, because that path stamps m_clientGUIDs and the
    crew must stay out of it -- they ride the map-membership channel, and the
    elimination sweep destroys whatever it does not re-find.
  * the viewer-side relay: Camera collects RelaySources and runs the same
    notifier over the far side, with ONE elimination over both.

AnnounceVessel/RetractVessel carry the hull AND THE CREW ONLY. Players and pets
are ordinary objects in ordinary cells on the deck and are found by the ordinary
sweep. Crew first on retract, hull last, or the client loses the ship while it
still holds them.

LOGIN, LOGOUT AND THE WIRE. The client is never told a deck map id -- it has no
terrain for one and dies in CMap::LoadWdt(). Player::GetClientMapId() is the map
the client is rendering (the water she sails) and every update packet is stamped
from it; SMSG_UPDATE_OBJECT carries it as a uint16, which a minted id would not
survive anyway. Player::BoardingMap() is the door at login and at the world-port
ack. PlayerSave writes the map the ship SAILS: written once, a deck id in that
column is unloadable for ever and the character never gets past the loading
screen.

DEATH. No body is ever left on a ship -- she is a map that sails away and whose
grids are pinned for the life of the server, so a corpse aboard is one nobody can
walk back to. GetWorldAnchor() answers with the vessel's pose on the map she
sails, so the nearest graveyard to it IS the port she is off, which is where
RepopAtGraveyard revives him. AnchorTerrain() keeps zone and area questions off
the hull, which carries no area table.

And the one that kills a server rather than a client: MapPersistentState took a
uint16 map id. TransportMap derives from Map, so a deck gets a persistent state
with its minted id -- 1020808 & 0xFFFF = 37768, GetMapEntry() finds nothing, and
every caller that dereferences it crashes. KILLING ANYTHING ON A DECK IS ENOUGH,
because the respawn bookkeeping runs through it. Widened here with CreatureData
and GameObjectData, following 00 -- the only core that had it right. Read, not
assumed: 01, 02 and the mangos_four reference all still carry uint16, although
TRANSPORTS.md claims 02 was widened. The code is the source.

Three defects found while wiring, all from the same recipe:

  * ~MapManager deleted the maps before the vessels, and crew unregister from
    their map's store as they are torn down -- that order reads freed memory.
  * UnloadAll never touched vessels at all. Nothing else frees one: a vessel is
    in no cell, so no grid unload reaches it.
  * SendInitTransports keyed on player->GetMapId() rather than i_id (not the same
    map during a teleport) and used tmap[...] -- a std::map subscript, which
    INSERTS an empty set into a container every map on the server shares.

5.4.8 adaptations, none of them cosmetic: no MOVEFLAG_ONTRANSPORT (being aboard
is carried by the transport data itself); UpdateData is map-scoped and
BuildPacket takes no hasTransport byte (that field is CLASSIC/TBC only);
MapEntry::ID and DBCString as a locale array; TaxiPathNodeEntry x/y/z; the hull's
update flags are TRANSPORT|HAS_POSITION|ROTATION with no LOWGUID/HIGHGUID. And
the create block now sends GetPathProgress() rather than the wall clock -- fed a
raw clock, the client stops animating the hull altogether.

vessels.txt gains the map id, and is no longer empty. Its header records what was
measured rather than assumed, from a real 5.4.8 client and a four_world database:

  * the Transport* rows in Map.dbc carry NO terrain -- the whole 30-archive chain
    has no `world\maps\Transport*` entry, and all 518 transport-shaped files are
    .wmo models. They are pseudo-maps; the deck is baked from the hull model into
    THE CLIENT'S OWN map id, never into one the server invents.
  * entry -> displayId is nowhere in the client. GameObjects.db2 (1780 rows,
    build 18273) holds no transport at all; TransportAnimation.dbc keys 113 lifts
    and platforms, none of these vessels; TransportRotation.dbc likewise.
  * but the baker does not need the entry -- mapId -> displayId is enough, and
    where the directory gives no entry the hull is still findable by model path.
  * 45 of the 49 maps now carry map, model and entry.

Known gaps, left marked rather than guessed: four maps still want an entry or a
model, "Transport197349-2" and "Transport197349" both parse to an entry that
exists nowhere, map 592 is a test row, and two directories carry a seven-digit
typo. 186371 "Stolen Zeppelin" is the one vessel that genuinely has no client
map and so genuinely needs a minted id.

NOT DONE, and it is data rather than code: `transports` holds 35 Cataclysm-era
rows against the client's 49 vessel maps, and 21 vessels have no route period.
Five of those need only a `transports` row -- they already have a
gameobject_template with a taxi path (176244/293, 177233/777, 201834/1834,
203729/2214, 203626/2265).

+++ src/game/WorldHandlers/MapPersistentStateMgr.h
@@ -65,7 +65,14 @@ class MapPersistentState
 {
         friend class MapPersistentStateManager;
     protected:
-        MapPersistentState(uint16 MapId, uint32 InstanceId, Difficulty difficulty);
+        // uint32, NOT uint16. A vessel's deck is a map like any other and gets a
+        // persistent state in the Map constructor -- with a MINTED id of
+        // VESSEL_MAP_BASE + goEntry. 1020808 & 0xFFFF = 37768, GetMapEntry() finds
+        // nothing under that, and every caller that dereferences it crashes. Killing
+        // anything on a deck is enough to reach it: the respawn bookkeeping runs through
+        // the persistent state. 01, 02 and this core all carried the narrow field; 00 is
+        // the one that widened it.
+        MapPersistentState(uint32 MapId, uint32 InstanceId, Difficulty difficulty);
 
     public:
 
@@ -161,7 +168,7 @@ class WorldPersistentState : public MapPersistentState
            - any new non-instanceable map created
            - respawn data loading for non-instanceable map
         */
-        explicit WorldPersistentState(uint16 MapId) : MapPersistentState(MapId, 0, REGULAR_DIFFICULTY) {}
+        explicit WorldPersistentState(uint32 MapId) : MapPersistentState(MapId, 0, REGULAR_DIFFICULTY) {}
 
         ~WorldPersistentState() {}
 
@@ -189,7 +196,7 @@ class DungeonPersistentState : public MapPersistentState
            - any new instance is being generated
            - the first time a player bound to InstanceId logs in
            - when a group bound to the instance is loaded */
-        DungeonPersistentState(uint16 MapId, uint32 InstanceId, Difficulty difficulty, time_t resetTime, bool canReset, uint32 completedEncountersMask);
+        DungeonPersistentState(uint32 MapId, uint32 InstanceId, Difficulty difficulty, time_t resetTime, bool canReset, uint32 completedEncountersMask);
 
         ~DungeonPersistentState();
 
@@ -263,7 +270,7 @@ class BattleGroundPersistentState : public MapPersistentState
         /* Created either when:
            - any new BG/arena is being generated
         */
-        BattleGroundPersistentState(uint16 MapId, uint32 InstanceId, Difficulty difficulty)
+        BattleGroundPersistentState(uint32 MapId, uint32 InstanceId, Difficulty difficulty)
             : MapPersistentState(MapId, InstanceId, difficulty) {}
 
         ~BattleGroundPersistentState() {}
